+++ dodam-components/src/main/java/kr/hs/dgsw/smartschool/components/component/organization/card/DodamCard.kt
@@ -4,6 +4,7 @@ import androidx.compose.foundation.layout.Box
 import androidx.compose.foundation.layout.Column
 import androidx.compose.foundation.layout.Row
 import androidx.compose.foundation.layout.Spacer
+import androidx.compose.foundation.layout.defaultMinSize
 import androidx.compose.foundation.layout.fillMaxWidth
 import androidx.compose.foundation.layout.height
 import androidx.compose.foundation.layout.padding
@@ -19,14 +20,18 @@ import androidx.compose.ui.graphics.Shape
 import androidx.compose.ui.text.TextStyle
 import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.dp
+import kr.hs.dgsw.smartschool.components.component.basic.badge.DodamBadge
 import kr.hs.dgsw.smartschool.components.foundation.Text
 import kr.hs.dgsw.smartschool.components.theme.Body3
 import kr.hs.dgsw.smartschool.components.theme.DodamTheme
+import kr.hs.dgsw.smartschool.components.theme.IcBreakfast3D
 import kr.hs.dgsw.smartschool.components.theme.IcDinner3D
+import kr.hs.dgsw.smartschool.components.theme.IcLunch3D
 import kr.hs.dgsw.smartschool.components.theme.IcOut3D
 import kr.hs.dgsw.smartschool.components.theme.IcRightArrow
 import kr.hs.dgsw.smartschool.components.theme.contentColorFor
 import kr.hs.dgsw.smartschool.components.utlis.DodamDimen
+import kr.hs.dgsw.smartschool.components.utlis.DodamString
 
 /**
  * Dodam Item Card, set title, subtitle, icon
@@ -163,6 +168,107 @@ fun DodamContentCard(
         }
     }
 }
+
+sealed interface MealType {
+    object BreakFast : MealType
+    object Lunch : MealType
+    object Dinner : MealType
+}
+
+private val MEAL_ICON_SIZE = 30.dp
+
+/**
+ * Dodam meal card
+ *
+ * @param content meal content
+ * @param mealType type of meal
+ * @param modifier modifier
+ * @param onClick action when click card
+ * @param shape shape of card
+ * @param background background color of card
+ * @param rippleColor rippleColor
+ * @param rippleEnable rippleEnable
+ * @param bounded bounded
+ */
+@Composable
+fun DodamMealCard(
+    content: String,
+    mealType: MealType,
+    modifier: Modifier = Modifier,
+    onClick: (() -> Unit)? = null,
+    shape: Shape = DodamTheme.shape.large,
+    background: Color = DodamTheme.color.White,
+    rippleColor: Color = Color.Unspecified,
+    rippleEnable: Boolean = true,
+    bounded: Boolean = true,
+) {
+
+    val mealData: MealData = when (mealType) {
+        MealType.BreakFast -> MealData(DodamString.Breakfast, DodamTheme.color.Breakfast) {
+            IcBreakfast3D(
+                modifier = Modifier.size(MEAL_ICON_SIZE),
+                contentDescription = null
+            )
+        }
+        MealType.Lunch -> MealData(DodamString.Lunch, DodamTheme.color.Lunch) {
+            IcLunch3D(
+                modifier = Modifier.size(MEAL_ICON_SIZE),
+                contentDescription = null
+            )
+        }
+        MealType.Dinner -> MealData(DodamString.Dinner, DodamTheme.color.Dinner) {
+            IcDinner3D(
+                modifier = Modifier.size(MEAL_ICON_SIZE),
+                contentDescription = null
+            )
+        }
+    }
+
+    Box {
+        Card(
+            onClick = onClick,
+            modifier = modifier
+                .fillMaxWidth()
+                .defaultMinSize(minHeight = 80.dp)
+                .padding(top = 15.dp),
+            shape = shape,
+            background = background,
+            rippleEnable = rippleEnable,
+            rippleColor = rippleColor,
+            bounded = bounded,
+        ) {
+            Row(
+                modifier = Modifier
+                    .padding(
+                        horizontal = DodamDimen.CardSidePadding,
+                        vertical = 24.dp
+                    )
+                    .fillMaxWidth()
+            ) {
+                DodamBadge(
+                    text = mealData.name,
+                    background = mealData.badgeColor,
+                    modifier = Modifier.align(Alignment.CenterVertically),
+                )
+                Spacer(modifier = Modifier.width(DodamDimen.CardSidePadding))
+                Text(
+                    text = content,
+                    style = DodamTheme.typography.body3,
+                    modifier = Modifier.align(Alignment.CenterVertically),
+                )
+            }
+        }
+        Box(modifier = Modifier.padding(start = 24.dp)) {
+            mealData.icon()
+        }
+    }
+}
+
+private data class MealData(
+    val name: String,
+    val badgeColor: Color,
+    val icon: @Composable () -> Unit
+)
 
 @Preview(showBackground = true)
 @Composable
